+++ app/Modules/Messages/Http/Controllers/Admin/TypesController.php
@@ -122,7 +122,7 @@ class TypesController extends Controller
 
 		$id = $request->input('id');
 
-		$row = $id ? Type::findOrFail($id) : new Type();
+		$row = Type::findOrNew($id);
 
 		$row->fill($request->input('fields'));
 
@@ -149,7 +149,12 @@ class TypesController extends Controller
 
 		foreach ($ids as $id)
 		{
-			$row = Type::findOrFail($id);
+			$row = Type::find($id);
+
+			if (!$row)
+			{
+				continue;
+			}
 
 			if (!$row->delete())
 			{
