+++ app/Modules/News/Http/Controllers/Admin/ArticlesController.php
@@ -450,7 +450,12 @@ class ArticlesController extends Controller
 		{
 			// Delete the entry
 			// Note: This is recursive and will also remove all descendents
-			$row = Article::findOrFail($id);
+			$row = Article::find($id);
+
+			if (!$row)
+			{
+				continue;
+			}
 
 			if (!$row->delete())
 			{
