+++ app/Modules/News/Http/Controllers/Admin/UpdatesController.php
@@ -205,7 +205,12 @@ class UpdatesController extends Controller
 		{
 			// Delete the entry
 			// Note: This is recursive and will also remove all descendents
-			$row = Update::findOrFail($id);
+			$row = Update::find($id);
+
+			if (!$row)
+			{
+				continue;
+			}
 
 			if (!$row->delete())
 			{
