+++ app/Modules/Queues/Http/Controllers/Admin/SchedulersController.php
@@ -113,15 +113,16 @@ class SchedulersController extends Controller
 
 	/**
 	 * Show the form for creating a new queue.
-	 * 
+	 *
+	 * @param  Request $request
 	 * @return View
 	 */
-	public function create()
+	public function create(Request $request)
 	{
 		$row = new Scheduler();
 		$row->defaultmaxwalltime = 1209600;
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -141,14 +142,15 @@ class SchedulersController extends Controller
 	/**
 	 * Show the form for editing the specified queue.
 	 *
+	 * @param  Request $request
 	 * @param  int $id
 	 * @return View
 	 */
-	public function edit($id)
+	public function edit(Request $request, $id)
 	{
 		$row = Scheduler::findOrFail($id);
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -193,7 +195,7 @@ class SchedulersController extends Controller
 
 		$id = $request->input('id');
 
-		$row = $id ? Scheduler::findOrFail($id) : new Scheduler();
+		$row = Scheduler::findOrNew($id);
 		$row->fill($request->input('fields'));
 
 		$walltime = $request->input('maxwalltime');
@@ -246,6 +248,11 @@ class SchedulersController extends Controller
 		foreach ($ids as $id)
 		{
 			$row = Scheduler::findOrFail($id);
+
+			if (!$row)
+			{
+				continue;
+			}
 
 			if (!$row->delete())
 			{
