+++ app/Modules/Tags/Http/Controllers/Admin/TaggedController.php
@@ -85,13 +85,14 @@ class TaggedController extends Controller
 	/**
 	 * Show the form for creating a new resource.
 	 *
+	 * @param  Request $request
 	 * @return View
 	 */
-	public function create()
+	public function create(Request $request)
 	{
 		$row = new Tagged();
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -145,14 +146,15 @@ class TaggedController extends Controller
 	/**
 	 * Show the form for editing the specified resource.
 	 *
+	 * @param  Request $request
 	 * @param  int $id
 	 * @return View
 	 */
-	public function edit($id)
+	public function edit(Request $request, $id)
 	{
 		$row = Tagged::findOrFail($id);
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -177,7 +179,12 @@ class TaggedController extends Controller
 
 		foreach ($ids as $id)
 		{
-			$row = Tagged::findOrFail($id);
+			$row = Tagged::find($id);
+
+			if (!$row)
+			{
+				continue;
+			}
 
 			if (!$row->delete())
 			{
