+++ app/Modules/Users/Http/Controllers/Admin/LevelsController.php
@@ -82,10 +82,11 @@ class LevelsController extends Controller
 
 	/**
 	 * Show the form for creating a new resource.
-	 * 
+	 *
+	 * @param  Request $request
 	 * @return View
 	 */
-	public function create()
+	public function create(Request $request)
 	{
 		$row = new Level;
 
@@ -93,7 +94,7 @@ class LevelsController extends Controller
 			->orderBy('lft', 'asc')
 			->get();
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -106,11 +107,12 @@ class LevelsController extends Controller
 
 	/**
 	 * Show the form for editing the specified resource.
-	 * 
+	 *
+	 * @param  Request $request
 	 * @param  int  $id
 	 * @return View
 	 */
-	public function edit($id)
+	public function edit(Request $request, $id)
 	{
 		$row = Level::findOrFail($id);
 
@@ -118,7 +120,7 @@ class LevelsController extends Controller
 			->orderBy('lft', 'asc')
 			->get();
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -180,7 +182,12 @@ class LevelsController extends Controller
 
 		foreach ($ids as $id)
 		{
-			$row = Level::findOrFail($id);
+			$row = Level::find($id);
+
+			if (!$row)
+			{
+				continue;
+			}
 
 			if (!$row->delete())
 			{
