+++ app/Modules/Users/Http/Controllers/Admin/RolesController.php
@@ -60,10 +60,11 @@ class RolesController extends Controller
 
 	/**
 	 * Show the form for creating a new resource.
-	 * 
+	 *
+	 * @param  Request $request
 	 * @return View
 	 */
-	public function create()
+	public function create(Request $request)
 	{
 		$row = new Role;
 
@@ -71,7 +72,7 @@ class RolesController extends Controller
 			->orderBy('lft', 'asc')
 			->get();
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -177,11 +178,12 @@ class RolesController extends Controller
 
 	/**
 	 * Show the form for editing the specified resource.
-	 * 
+	 *
+	 * @param  Request $request
 	 * @param  int  $id
 	 * @return View
 	 */
-	public function edit($id)
+	public function edit(Request $request, $id)
 	{
 		$row = Role::findOrFail($id);
 
@@ -190,7 +192,7 @@ class RolesController extends Controller
 			->orderBy('lft', 'asc')
 			->get();
 
-		if ($fields = app('request')->old('fields'))
+		if ($fields = $request->old('fields'))
 		{
 			$row->fill($fields);
 		}
@@ -218,6 +220,11 @@ class RolesController extends Controller
 		foreach ($ids as $id)
 		{
 			$row = Role::findOrFail($id);
+
+			if (!$row)
+			{
+				continue;
+			}
 
 			if (!$row->delete())
 			{
